支付价格获取

+++ hltrip_central_web/src/main/java/com/huoli/trip/central/web/service/impl/DfyOrderManager.java
@@ -353,7 +353,7 @@ public class DfyOrderManager extends OrderManager {
         request.setTraceId(req.getTraceId());
         request.setChannelCode(req.getChannelCode());
         request.setChannelOrderId(req.getChannelOrderId());
-        request.setPrice(String.valueOf(req.getPrice()));
+        request.setPrice(String.valueOf(req.getPrice().intValue()));
         DfyBaseResult dfyBaseResult = dfyOrderService.payOrder(request);
         if(dfyBaseResult != null && dfyBaseResult.isSuccess()){
             CenterPayOrderRes payOrderRes = new CenterPayOrderRes();
